Change behavior when deleting a profile -- instead of pausing jobs, you must now resolve conflicts before deleting the profile. Also fix typo in "STYLES_PATH"

+++ admin/views.py
@@ -7,10 +7,11 @@ import os
 from datetime import datetime, timedelta
 from apscheduler.triggers.cron import CronTrigger
 from urllib.parse import urlparse, urljoin
-from . import load_profiles, Permissions, DATABASE_PATH, scheduler, jobstores
+from . import load_profiles, Permissions, DATABASE_PATH, STYLES_PATH, scheduler, jobstores
 import json
 import copy
 import shutil
+import queue
 
 admin = Blueprint('admin', __name__, template_folder="templates")
 
@@ -414,20 +415,24 @@ def edit_profile(conn):
   conn.close()
 
 def delete_profile(conn):
-  cursor = conn.cursor()
-
   profile_name = request.form['profile_name']
+  msg_stack = queue.LifoQueue()
+  for job in scheduler.get_jobs(jobstore='league'):
+    if profile_name in job.args:
+      msg_stack.put("Cannot delete profile; it is in use by league job ID: {:s}".format(job.id))
+
+  if not msg_stack.empty():
+    while not msg_stack.empty():
+      flash(msg_stack.get())
+    conn.close()
+    return
+
+  cursor = conn.cursor()
   cursor.execute("SELECT name FROM server_profiles WHERE name=?", (profile_name,))
 
   if cursor.fetchone():
     cursor.execute("DELETE FROM server_profiles WHERE name=?", (profile_name,))
     conn.commit()
-
-    for job in scheduler.get_jobs(jobstore='league'):
-      if profile_name in job.args:
-        scheduler.pause_job(job.id, jobstore='league')
-        flash('Profile was in use by some scheduled league jobs. Those jobs have been paused.')
-
     flash('Profile deleted successfully!')
   else:
     flash('Profile not found!')
@@ -680,11 +685,10 @@ def upload_styles():
     style_name = style_name.strip().replace(" ", "_").lower()
     file_data = request.files['style_file']
     if file_data:
-      styles_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'static', 'app_styles')
-      if not os.path.exists(styles_dir):
-        os.makedirs(styles_dir)
+      if not os.path.exists(STYLES_PATH):
+        os.makedirs(STYLES_PATH)
 
-      file_path = os.path.join(styles_dir, "{:s}.json".format(style_name))
+      file_path = os.path.join(STYLES_PATH, "{:s}.json".format(style_name))
       style_json = file_data.read().decode('utf-8')
       style = json.loads(style_json)
 
@@ -706,17 +710,16 @@ def change_styles():
   if not check_permissions(Permissions.CHANGE_STYLES):
     return redirect(url_for('admin.unauthorized'))
 
-  styles_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'static', 'app_styles')
-  style_files = os.listdir(styles_dir) if os.path.exists(styles_dir) else []
-  style_files = filter(lambda f: f.endswith('.json') and not os.path.islink(os.path.join(styles_dir, f)), style_files)
+  style_files = os.listdir(STYLES_PATH) if os.path.exists(STYLES_PATH) else []
+  style_files = filter(lambda f: f.endswith('.json') and not os.path.islink(os.path.join(STYLES_PATH, f)), style_files)
 
   if request.method == 'POST':
     selected_style = request.form['style_file']
 
-    file_path = os.path.join(styles_dir, "{:s}.json".format(selected_style))
+    file_path = os.path.join(STYLES_PATH, "{:s}.json".format(selected_style))
     tmp_name = "{:s}_tmp.json".format(str(uuid.uuid4()))
-    os.symlink(file_path, os.path.join(styles_dir, tmp_name))
-    os.replace(os.path.join(styles_dir, tmp_name), os.path.join(styles_dir, "user_styles.json"))
+    os.symlink(file_path, os.path.join(STYLES_PATH, tmp_name))
+    os.replace(os.path.join(STYLES_PATH, tmp_name), os.path.join(STYLES_PATH, "user_styles.json"))
     flash('Style changed successfully!')
 
   current_style = "default_styles"
@@ -732,10 +735,9 @@ def delete_style():
   if not check_permissions(Permissions.CHANGE_STYLES):
     return redirect(url_for('admin.unauthorized'))
 
-  styles_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'static', 'app_styles')
   style_name = request.form['style_file']
-  file_path = os.path.join(styles_dir, "{:s}.json".format(style_name))
-  recycle_bin_path = os.path.join(styles_dir, "recycle_bin")
+  file_path = os.path.join(STYLES_PATH, "{:s}.json".format(style_name))
+  recycle_bin_path = os.path.join(STYLES_PATH, "recycle_bin")
   if not os.path.exists(recycle_bin_path):
     os.makedirs(recycle_bin_path)
